requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-aiplatform